Move pipeline status into sub-app
