+++ tests/integration/test_modal_pipeline_seams.py
@@ -56,6 +56,7 @@ def test_pipeline_image_runtime_seams():
         "uv.lock": True,
         "modal_app/worker_script.py": True,
         "modal_app/local_area.py": True,
+        "modal_app/pipeline_status.py": True,
         "modal_app/h5_test_harness.py": True,
         "modal_app/step_manifests/specs.py": True,
         "modal_app/step_manifests/state.py": True,
@@ -78,6 +79,7 @@ def test_pipeline_image_runtime_seams():
         "modal_app.fixtures.h5_cases",
         "modal_app.h5_test_harness",
         "modal_app.local_area",
+        "modal_app.pipeline_status",
         "modal_app.remote_calibration_runner",
         "modal_app.step_manifests.specs",
         "modal_app.step_manifests.state",
